- Fixed Memory Bug in block_collection - Jpeg Classification Workaround: Removed FT_HIGH_ENTROPY from JPEG
git-svn-id: https://showcase.fhstp.ac.at/republic/mmc/trunk@901 76e0c08c-8e14-4548-9514-92ecd803f47c

+++ collating/fragment/include/block_collection.h
@@ -7,7 +7,7 @@
 #define HEADER(x) ((x) & 0x02)
 
 #define BITS_PER_BLOCK 2
-#define BLOCKS_PER_STORAGE (sizeof(unsigned long) * 8/ BITS_PER_BLOCK)
+#define BLOCKS_PER_STORAGE (sizeof(unsigned long long) * 8/ BITS_PER_BLOCK)
 
 typedef unsigned long long storage_t;
 typedef struct _block_collection_t block_collection_t;
